Integrate region streaming into runepy.world
