--- world/__init__.py
@@ -1,6 +1,0 @@
-from .coords import world_to_region, local_tile
-from .region import Region
-from .region_manager import RegionManager
-
-from .world import World
-__all__ = ["Region", "RegionManager", "World", "world_to_region", "local_tile"]
